# openapi.yml
openapi: 3.1.0
info:
  title: MemoryOS browser API
  description: The same-origin contract consumed by the MemoryOS browser application.
  version: 1.0.0
servers:
- url: /
  description: Same-origin MemoryOS runtime
paths:
...
  /api/identity/me/language:
    put:
      tags:
      - Identity
      summary: Set the authenticated account's interface language
      operationId: setCurrentIdentityLanguage
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LanguagePreference"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/LanguagePreference"
      security:
      - browserSession: []
      - bearerAuth: []
...
  /api/credentials/google-drive/{credentialId}/revoke:
    post:
      tags:
      - Credentials
      summary: Revoke a shared Google Drive credential and disconnect all attached
        Sources
      operationId: revokeGoogleDriveCredential
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: credentialId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeGoogleDriveCredentialRequest"
        required: true
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
      - browserSession: []
  /api/credentials/google-drive/authorization:
    post:
      tags:
      - Credentials
      summary: Authorize a reusable Google Drive credential
      operationId: startGoogleDriveAuthorization
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartGoogleDriveAuthorizationRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/GoogleDriveAuthorizationResponse"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/sources/{sourceId}:
    get:
      tags:
      - Sources
      summary: Get one source summary
      operationId: getSource
      parameters:
      - name: sourceId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SourceSummary"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/identity/me:
    get:
      tags:
      - Identity
      summary: Return the authenticated MemoryOS actor
      description: Accepts either an existing MemoryOS browser session or a valid
        bound bearer identity and returns the stable internal ActorId plus its durable
        Tenant authority projection.
      operationId: getCurrentIdentity
      responses:
        "200":
          description: The authenticated actor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentIdentity"
        "401":
          description: No accepted authentication is present
      security:
      - browserSession: []
      - bearerAuth: []
...
  /api/credentials/google-drive:
    get:
      tags:
      - Credentials
      summary: List reusable Tenant-owned Google Drive credentials
      operationId: listGoogleDriveCredentials
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GoogleDriveCredentialResponse"
      security:
      - bearerAuth: []
      - browserSession: []
...
  /api/credentials/google-drive/{credentialId}:
    delete:
      tags:
      - Credentials
      summary: Delete an unused Google Drive credential with a revision precondition
      operationId: deleteGoogleDriveCredential
      parameters:
      - name: X-MemoryOS-CSRF
        in: header
        description: Same-origin non-simple request guard for browser-session mutations.
        required: true
        schema:
          type: string
          enum:
          - "1"
      - name: credentialId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: If-Match
        in: header
        required: true
        schema:
          type: string
      responses:
        "204":
          description: No Content
      security:
      - bearerAuth: []
      - browserSession: []
components:
  schemas:
...
    LanguagePreference:
      type: object
      additionalProperties: false
      properties:
        uiLanguage:
          type: string
          enum:
          - vi
          - en
          pattern: vi|en
      required:
      - uiLanguage
...
    SourceSummary:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        access:
          type: string
        status:
          type: string
        pendingWork:
          type: boolean
        documentCount:
          type: integer
          format: int64
        lastSucceededAt:
          type:
          - string
          - "null"
          format: date-time
        errorCode:
          type:
          - string
          - "null"
        actions:
          type: array
          items:
            type: string
            enum:
            - upload
            - reindex
            - remove_items
            - delete
            - manage_groups
            - rename
            - manage_access
            - manage_configuration
            - synchronize
            - manage_schedule
            - pause_sync
            - resume_sync
      required:
      - access
      - actions
      - documentCount
      - errorCode
      - id
      - lastSucceededAt
      - name
      - pendingWork
      - status
      - type
...
    RevokeGoogleDriveCredentialRequest:
      type: object
      additionalProperties: false
      properties:
        expectedCredentialRevision:
          type: integer
          format: int64
          minimum: 0
      required:
      - expectedCredentialRevision
    StartGoogleDriveAuthorizationRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 120
          minLength: 0
        credentialId:
          type: string
          format: uuid
        expectedCredentialRevision:
          type: integer
          format: int64
          minimum: 0
        oauthClientJson:
          type: string
          maxLength: 16384
          minLength: 0
      required:
      - name
    GoogleDriveAuthorizationResponse:
      type: object
      additionalProperties: false
      properties:
        authorizationUrl:
          type: string
      required:
      - authorizationUrl
...
    CurrentIdentity:
      type: object
      additionalProperties: false
      properties:
        actorId:
          type: string
          format: uuid
          description: Stable internal MemoryOS actor identifier.
        tenant:
          description: "Active Tenant context, or null when the actor has no active\
            \ Tenant membership."
          oneOf:
          - $ref: "#/components/schemas/CurrentTenant"
          - type: "null"
        capabilities:
          type: array
          description: Expanded global capabilities backed by current server enforcement.
          items:
            type: string
            enum:
            - SYSTEM_ADMIN
            - SYSTEM_BASIC
            - SEARCH_READ
            - CHAT_READ
            - CHAT_WRITE
            - IMAGE_GENERATE
            - LLM_GATEWAY_USE
            - USERS_MANAGE
            - GROUPS_READ
            - GROUPS_MANAGE
            - SOURCES_READ
            - SOURCES_MANAGE
            - SOURCES_DELETE
            - MODELS_MANAGE
        scopedCapabilities:
          type: array
          description: Eligible capabilities available only within resources managed
            by this actor.
          items:
            type: string
            enum:
            - SYSTEM_ADMIN
            - SYSTEM_BASIC
            - SEARCH_READ
            - CHAT_READ
            - CHAT_WRITE
            - IMAGE_GENERATE
            - LLM_GATEWAY_USE
            - USERS_MANAGE
            - GROUPS_READ
            - GROUPS_MANAGE
            - SOURCES_READ
            - SOURCES_MANAGE
            - SOURCES_DELETE
            - MODELS_MANAGE
        authorizationVersion:
          type: integer
          format: int64
          description: Monotonic Tenant IAM revision used only to invalidate private
            client data.
          minimum: 0
        uiLanguage:
          type: string
          description: Account interface language.
          enum:
          - vi
          - en
      required:
      - actorId
      - authorizationVersion
      - capabilities
      - scopedCapabilities
      - tenant
      - uiLanguage
    CurrentTenant:
      type: object
      additionalProperties: false
      properties:
        displayName:
          type: string
          description: Display name of the actor's active Tenant.
        role:
          type: string
          description: Stable Tenant membership role used for presentation.
          enum:
          - OWNER
          - MEMBER
      required:
      - displayName
      - role
...
    GoogleDriveCredentialResponse:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        accountEmail:
          type: string
        status:
          type: string
        credentialRevision:
          type: integer
          format: int64
        oauthClientConfigured:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        sourceCount:
          type: integer
          format: int64
        actions:
          type: array
          items:
            type: string
      required:
      - accountEmail
      - actions
      - createdAt
      - credentialRevision
      - id
      - name
      - oauthClientConfigured
      - sourceCount
      - status
      - updatedAt
...
  securitySchemes:
    browserSession:
      type: apiKey
      description: HttpOnly JDBC-backed Spring Security session cookie.
      name: SESSION
      in: cookie
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT